feat: change way indent and quotes are handled by conversion schematics

+++ packages/schematics/tests/convert-tslint-to-eslint/convert-to-eslint-config.test.ts
@@ -182,7 +182,7 @@ describe('convertToESLintConfig()', () => {
             "@typescript-eslint/prefer-namespace-keyword": "error",
             "@typescript-eslint/quotes": Array [
               "error",
-              "single",
+              "double",
             ],
             "@typescript-eslint/semi": Array [
               "error",
